OLE-9004 : Refactored the code. Merged the code from 2.1 to 3.0 Code base.
