+++ ole-app/olefs/src/main/java/org/kuali/ole/oleng/dao/export/ExportBibIdFinderCallable.java
@@ -1,14 +1,8 @@
 package org.kuali.ole.oleng.dao.export;
 
-import org.apache.commons.lang.StringUtils;
-import org.apache.solr.common.SolrDocument;
-import org.apache.solr.common.SolrDocumentList;
-import org.kuali.ole.DocumentUniqueIDPrefix;
-import org.kuali.ole.constants.OleNGConstants;
 import org.kuali.ole.oleng.handler.BatchExportHandler;
+import org.kuali.ole.oleng.util.BatchExportUtil;
 
-import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
 import java.util.concurrent.Callable;
 
@@ -21,6 +15,7 @@ public class ExportBibIdFinderCallable implements Callable {
     private int start;
     private int chunkSize;
     private BatchExportHandler batchExportHandler;
+    private BatchExportUtil batchExportUtil;
 
     public ExportBibIdFinderCallable(String query, int start, int chunkSize, BatchExportHandler batchExportHandler) {
         this.query = query;
@@ -32,18 +27,18 @@ public class ExportBibIdFinderCallable implements Callable {
 
     @Override
     public Object call() throws Exception {
-        Set<String> bibIdentifiers = new HashSet<>();
-        SolrDocumentList solrDocumentList = batchExportHandler.getSolrRequestReponseHandler().getSolrDocumentList(query, start, chunkSize, OleNGConstants.BIB_IDENTIFIER);
-        if (solrDocumentList.size() > 0) {
-            for (SolrDocument solrDocument : solrDocumentList) {
-                if (solrDocument.containsKey(OleNGConstants.BIB_IDENTIFIER)) {
-                    List<String> bibIds = (List) solrDocument.getFieldValue(OleNGConstants.BIB_IDENTIFIER);
-                    for (String bibId : bibIds) {
-                        bibIdentifiers.add(DocumentUniqueIDPrefix.getDocumentId(bibId));
-                    }
-                }
-            }
-        }
+        Set<String> bibIdentifiers = getBatchExportUtil().getBibIdentifiersForQuery(query, start, chunkSize);
         return bibIdentifiers;
+    }
+
+    public BatchExportUtil getBatchExportUtil() {
+        if(null == batchExportUtil) {
+            batchExportUtil = new BatchExportUtil();
+        }
+        return batchExportUtil;
+    }
+
+    public void setBatchExportUtil(BatchExportUtil batchExportUtil) {
+        this.batchExportUtil = batchExportUtil;
     }
 }
