+++ ole-app/olefs/src/main/java/org/kuali/ole/oleng/handler/BatchExportHandler.java
@@ -44,31 +44,31 @@ public class BatchExportHandler extends BatchExportUtil {
 
     private void processFullExport(BatchProcessTxObject batchProcessTxObject, OleNGBatchExportResponse oleNGBatchExportResponse) {
         String query = "(DocType:bibliographic)";
-        exportDao.export(this, query, batchProcessTxObject, oleNGBatchExportResponse);
+        exportDao.export(this, query, batchProcessTxObject, oleNGBatchExportResponse, false);
     }
 
     private void processFullExceptStaffOnly(BatchProcessTxObject batchProcessTxObject, OleNGBatchExportResponse oleNGBatchExportResponse) {
         String query = "(DocType:bibliographic)AND(staffOnlyFlag:false)";
-        exportDao.export(this, query, batchProcessTxObject, oleNGBatchExportResponse);
+        exportDao.export(this, query, batchProcessTxObject, oleNGBatchExportResponse, false);
     }
 
     private void processIncremental(BatchProcessTxObject batchProcessTxObject, OleNGBatchExportResponse oleNGBatchExportResponse) {
         Date lastExportDate = getLastExportDateForProfile(batchProcessTxObject);
         String query = getIncrementalSolrQuery(lastExportDate);
-        exportDao.export(this, query, batchProcessTxObject, oleNGBatchExportResponse);
+        exportDao.export(this, query, batchProcessTxObject, oleNGBatchExportResponse, true);
         processDeletedAndStaffOnlyBibs(lastExportDate, batchProcessTxObject);
     }
 
     private void processIncrementalExceptStaffOnly(BatchProcessTxObject batchProcessTxObject, OleNGBatchExportResponse oleNGBatchExportResponse) {
         Date lastExportDateForProfile = getLastExportDateForProfile(batchProcessTxObject);
         String query = getIncrementalExceptStaffOnlySolrQuery(lastExportDateForProfile);
-        exportDao.export(this, query, batchProcessTxObject, oleNGBatchExportResponse);
+        exportDao.export(this, query, batchProcessTxObject, oleNGBatchExportResponse, true);
         processDeletedAndStaffOnlyBibs(lastExportDateForProfile, batchProcessTxObject);
     }
 
     private void processFilterExport(BatchProcessTxObject batchProcessTxObject, OleNGBatchExportResponse oleNGBatchExportResponse) {
         String query = getFilterSolrQuery(batchProcessTxObject, oleNGBatchExportResponse);
-        exportDao.export(this, query, batchProcessTxObject, oleNGBatchExportResponse);
+        exportDao.export(this, query, batchProcessTxObject, oleNGBatchExportResponse, false);
     }
 
 }
